Almost finished middle auto
